+++ src/vcstools/git.py
@@ -170,7 +170,7 @@ class GitClient(VcsClientBase):
             return False
 
         if refname is not None and refname != "master":
-            return self.update(refname)
+            return self.update(refname, verbose=verbose)
         else:
             return True
 
@@ -189,9 +189,15 @@ class GitClient(VcsClientBase):
         return True
 
     def update(self, refname=None, verbose=False):
-        """interprets refname as a local branch, remote branch, tagname, hash, etc.
-        If it is a branch, attempts to move to it unless already on it, and to fast-forward, unless not a tracking branch.
-        Else go untracked on tag or whatever refname is. Does not leave if current commit would become dangling."""
+        """
+        interprets refname as a local branch, remote branch, tagname,
+        hash, etc.
+
+        If it is a branch, attempts to move to it unless
+        already on it, and to fast-forward, unless not a tracking
+        branch. Else go untracked on tag or whatever refname is. Does
+        not leave if current commit would become dangling.
+        """
         # try calling git fetch just once per call to update()
         need_to_fetch = True
         if not self.detect_presence():
@@ -209,7 +215,7 @@ class GitClient(VcsClientBase):
             refname = branch_parent
         if refname is None:
             # we are neither tracking, nor did we get any refname to update to
-            return self.update_submodules()
+            return self.update_submodules(verbose=verbose)
 
         # local branch might be named differently from remote by user, we respect that
         same_branch = (refname == branch_parent) or (refname == current_branch)
@@ -235,7 +241,7 @@ class GitClient(VcsClientBase):
 
             # shortcut if version is the same as requested
             if not refname_is_branch and self.get_version() == refname:
-                return self.update_submodules()
+                return self.update_submodules(verbose=verbose)
 
             if current_branch is None:
                 current_version = self.get_version()
@@ -260,7 +266,7 @@ class GitClient(VcsClientBase):
                                                  verbose=verbose):
                         return False
 
-        return self.update_submodules()
+        return self.update_submodules(verbose=verbose)
 
     def get_version(self, spec=None):
         """
@@ -537,6 +543,8 @@ class GitClient(VcsClientBase):
         if parent is not None and self.rev_list_contains("remotes/origin/%s"%parent,
                                                          self.get_version(),
                                                          fetch=False):
+            if verbose:
+                print("Rebasing repository")
             # Rebase, do not pull, because somebody could have
             # commited in the meantime.
             if LooseVersion(self.gitversion) >= LooseVersion('1.7.1'):